PBI 24-25 Backend

+++ kanbanapi/src/main/java/sit/int221/kanbanapi/services/CollabService.java
@@ -48,7 +48,7 @@ public class CollabService {
         if (!isOwner) {
             throw new AuthenticationFailed("You do not have permission to perform this action.");
         }
-        if (isOwner && board.getOwnerId().equals(newCollab.getOid())) {
+        if (board.getOwnerId().equals(newCollab.getOid())) {
             throw new CollaboratorConflict("You are a board owner.");
         }
         Collab collab = new Collab();
@@ -62,18 +62,23 @@ public class CollabService {
 
     public Collab collabAccess(String boardId, String userOid, CollabAccessEditRequestDTO collabAccessEditRequestDTO) {
         Collab collab = collabRepository.findByBoardIdAndUserOid(boardId, userOid).orElseThrow(() -> new ItemNotFoundException("Collaborator not found"));
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new ItemNotFoundException("Board not found"));
+        User currentUser = userRepository.findByUsername(jwtUserDetailsService.getCurrentUser().getUsername());
+        if (currentUser.getOid().equals(board.getOwnerId())) {
+            throw new AuthenticationFailed("You do not have permission to perform this action.");
+        }
         collab.setAccessRight(collabAccessEditRequestDTO.getAccess_right());
         return collabRepository.save(collab);
     }
 
     public Collab deleteCollaborator(String boardId, String userOid) {
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new ItemNotFoundException("Board not found"));
         Collab collab = collabRepository.findByBoardIdAndUserOid(boardId, userOid).orElseThrow(() -> new ItemNotFoundException("Collaborator not found"));
+        User currentUser = userRepository.findByUsername(jwtUserDetailsService.getCurrentUser().getUsername());
+        if (currentUser.getOid().equals(board.getOwnerId()) || !currentUser.getOid().equals(userOid)) {
+            throw new AuthenticationFailed("You do not have permission to perform this action.");
+        }
         collabRepository.delete(collab);
         return collab;
-    }
-
-    public void checkCollaborator(String boardId, String userOid) {
-
-        User user = userRepository.findById(userOid).orElseThrow(() -> new ItemNotFoundException("Collaborator not found"));
     }
 }
